Created Jooq in program configuration
